requirements.txt
PyYaml == 5.3.1
psutil == 5.7.3
requests == 2.25.0
web.py == 0.62